\name{rexpar-package}
\alias{rexpar-package}
\alias{rexpar}
\docType{package}
\title{
rexpar - a robust test for explosive AR processes based on simplicial depth
}
\description{
The package contains simplicial depth statistics for autoregressive processes with explosion. Further tests
based on Leucht, Mueller and Kustosz (2014) are 
defined and some example functions are included. In future version we will also include estimators and evaluation functions, e.g. to examin the power of the proposed tests.
}
\details{
\tabular{ll}{
Package: \tab rexpar\cr
Type: \tab Package\cr
Version: \tab 0.99(1.0beta)\cr
Date: \tab 2014-07-25\cr
License: \tab GPL-2\cr
}
NOTE: THIS IS A PRELIMINARY VERSION OF THE PACKAGE !!! FOR INTERNAL USE ONLY !!! 
}
\author{
Kustosz, Christoph

Maintainer: Kustosz, Christoph <kustosz@statistik.tu-dortmund.de>
}
\references{
Kustosz, C. (2014). Implementation of simplicial depth for AR processes under non standard conditions. Working Paper, TU Dortmund.
Leucht, Kustosz and Mueller (2014). Tests based on simplicial depth for AR(1) models with
explosion. Submitted.
Kustosz, Mueller and Wendler (2014). Simplified Simplicial Depth for Regression and
Autoregressive Growth Processes. Submitted.
}

%\keyword{ package }
%\seealso{
%}
%\examples{
%}
