\name{rexpar-package}
\alias{rexpar-package}
\alias{rexpar}
\docType{package}
\title{
rexpar -  Robust Methods for Explosive Autoregressive Processes Based on Simplicial Depth
}
\description{
The package contains simplicial depth statistics for autoregressive processes with explosion. Tests statisticsbased on Kustosz and Müller (2014), Kustosz, Leucht and Müller (2016) and Kustosz, Müller and Wendler (2016) are implemented and some functions to generate examples are included. Further estimators, confidence sets are implementd. For the one dimensional AR Model, we also present a prediction algorithm and change point detection methods for one and two parameter models are proposed, as defined in Kustosz (2016). 
}
\details{
\tabular{ll}{
Package: \tab rexpar\cr
Type: \tab Package\cr
Version: \tab 0.99(1.0beta)\cr
Date: \tab 2014-07-25\cr
License: \tab GPL-2, ACM\cr
}
}
\author{
Christoph Kustosz and Sebastian Szugat

Maintainer: Kustosz, Christoph <kustosz@statistik.tu-dortmund.de>
}
\references{
Kustosz, C. (2016). Depth based estimators and tests for
autoregressive processes with application. Ph. D. thesis. TU Dortmund.\cr 
\cr
Kustosz C., Leucht A. and Müller Ch. H. (2016). Tests based on simplicial depth for AR(1) models with
explosion. Journal of Time Series Analysis. In press.\cr
\cr
Kustosz C., Müller Ch. H. and Wendler M. (2016). Simplified Simplicial Depth for Regression and
Autoregressive Growth Processes. Journal of Statistical Planning and Inference. In press.\cr
\cr
Kustosz C. and Müller Ch. H. (2014). Analysis of crack growth with robust distribution-
free estimators and tests for nonstationary autoregressive processes. Statistical
Papers, 55:125–140.
}

%\keyword{ package }
%\seealso{
%}
%\examples{
%}
